backlight/tests: cast fake-BAR len to u64 (fix kernel-test build)
The Wave fake-BAR resize (24684774) set MmioRegion.len = raw.len() (usize)
but the field is u64, an E0308 that only the test-module build (xtask test)
compiles — clippy on narf-frame never reaches it. Broke the kernel-test suite
build.

+++ drivers/backlight/src/tests.rs
@@ -429,7 +429,7 @@ mod smokes {
         // ~800 KiB stack frame.
         let buf: Box<[u8]> = alloc::vec![0u8; 0xC9000].into_boxed_slice();
         let raw = Box::leak(buf);
-        let bar_len = raw.len();
+        let bar_len = raw.len() as u64;
         let phys = narf_memory::PhysAddr::new(raw.as_ptr() as u64);
         let mmio = narf_driver_runtime::MmioRegion {
             phys,
